nettoyage code
